Until now to combine all possible codons in several positions, we used the carthesian product of all the sets of synonymous codons in 'righmost nesting order'. However this 'order', which is also the order in which combinations are tested (the first one matching constraints being the one choosed) was not optimal:
We want to avoid using too rare codons, so we put them at the end of
lists of synonymous codons.

But itertools.product will pick the combination (first of set one,
last of set two) before picking (second of set one, first of set two)
while the latter is often better.

Thus we now sort all the possible combinations according to a score
that is for now equal to max(order of choosed codon at pos k among
synonymous) + sum(order of choosed codon at pos k among synonymous).

This scoring is probably not optimal, and neither is the implementation
since it returns a list instead of an iterator.